add user types
