+++ admin-frontend/src/features/users/types.ts
@@ -1,0 +1,8 @@
+export interface User {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  createdAt?: string;
+  updatedAt?: string;
+}
